+++ autoload/omnipytent/integration/fzf.py
@@ -1,8 +1,11 @@
 import vim
 
+import sys
+
 from omnipytent.async_execution import SelectionUI
 from omnipytent.execution import FN, quote
 from omnipytent.util import RawVim
+from omnipytent import simple_tcp_loopback_server
 
 class FZF(SelectionUI):
     def gen_entry(self, i, item):
@@ -26,8 +29,20 @@ class FZF(SelectionUI):
         if self.prompt:
             flags.append('--prompt ' + quote(self.prompt))
 
+        if self.preview:
+            self.preview_server_cm = simple_tcp_loopback_server.socket_bound(self._bytes_for_preview)
+            preview_server_port = self.preview_server_cm.__enter__()
+            flags.append('--preview ' + quote('%s %s %d {1}' % (
+                quote(sys.executable),
+                quote(simple_tcp_loopback_server.__file__),
+                preview_server_port)))
+        else:
+            self.preview_server_cm = None
+
         params['options'] = ' '.join(map(str, flags))
         FN['fzf#run'](params)
 
     def finish(self, choice):
+        if self.preview_server_cm:
+            self.preview_server_cm.__exit__(None, None, None)
         self.run_next_frame('finish_indices', choice)